reorganzied pauli-fermi op stuff into runtime/utils
