+++ runtime/CMakeLists.txt
@@ -6,7 +6,7 @@ add_library(${LIBRARY_NAME} SHARED ${SRC})
 
 target_include_directories(${LIBRARY_NAME} PUBLIC . qrt)
 
-target_link_libraries(${LIBRARY_NAME} PUBLIC xacc::xacc xacc::quantum_gate qrt xacc::pauli xacc::fermion)
+target_link_libraries(${LIBRARY_NAME} PUBLIC xacc::xacc xacc::quantum_gate qrt utils)
 
 xacc_configure_library_rpath(${LIBRARY_NAME})
 
@@ -20,3 +20,4 @@ endif()
 
 add_subdirectory(objectives)
 add_subdirectory(qrt)
+add_subdirectory(utils)
